+++ vector/src/main/java/im/vector/activity/BugReportActivity.kt
@@ -76,6 +76,8 @@ class BugReportActivity : MXCActionBarActivity() {
             mScreenShotPreview.setImageBitmap(BugReporter.getScreenshot())
         } else {
             mScreenShotPreview.isVisible = false
+            mIncludeScreenShotButton.isChecked = false
+            mIncludeScreenShotButton.isEnabled = false
         }
     }